requirements.txt
streamlit>=1.22.0
torch>=2.0.1
opencv-python-headless>=4.7.0
boto3>=1.26.0
python-dotenv